+++ nn_dataflow/core/partition.py
@@ -19,6 +19,7 @@ program. If not, see <https://opensource.org/licenses/BSD-3-Clause>.
 """
 
 import itertools
+import fastcache
 
 from . import data_category_enum as de
 from . import parallel_enum as pe
@@ -242,6 +243,7 @@ def proc_data_range(layer, batch_size, part, pidx):
     return filrng, ifrng, ofrng
 
 
+@fastcache.clru_cache(maxsize=1024)
 def unit_nhops_to_proc_region(layer, batch_size, region, part,
                               filter_nodes, ifmap_layout, ofmap_layout,
                               options):
@@ -286,9 +288,33 @@ def unit_nhops_to_proc_region(layer, batch_size, region, part,
     assert len(set(len(v) for v in ofm_dict.values())) <= 1, \
             'ofm val len: {}'.format([len(v) for v in ofm_dict.values()])
 
+    fil_dict = util.HashableDict.fromdict(fil_dict, valfunc=tuple)
+    ifm_dict = util.HashableDict.fromdict(ifm_dict, valfunc=tuple)
+    ofm_dict = util.HashableDict.fromdict(ofm_dict, valfunc=tuple)
+
     nhops = [0] * de.NUM
 
-    # Filter access.
+    nhops[de.FIL] = _unit_nhops_to_fil(layer, filter_nodes, fil_dict)
+
+    nhops[de.IFM] = _unit_nhops_to_ifm(ifmap_layout, ifm_dict)
+
+    if ofmap_layout.parts == (part,) and ofmap_layout.regions == (region,):
+        # Ofmaps are stored locally, no data transfer.
+        pass
+    else:
+        nhops[de.OFM] = _unit_nhops_to_ofm(ofmap_layout, ofm_dict)
+
+    return nhops
+
+
+@fastcache.clru_cache(maxsize=1024)
+def _unit_nhops_to_fil(layer, filter_nodes, fil_dict):
+    '''
+    Get the total number of hops to transfer filter data.
+
+    `fil_dict` maps each filter range to the nodes that process it.
+    '''
+    nhops = 0
 
     for filrng, coord_list in fil_dict.items():
         fil_size = filrng[0].size() * filrng[1].size() * layer.filter_size()
@@ -296,14 +322,34 @@ def unit_nhops_to_proc_region(layer, batch_size, region, part,
         # Min hops to each processing node across all filter source nodes.
         min_hops = [min(coord.hop_dist(c) for c in filter_nodes)
                     for coord in coord_list]
-        nhops[de.FIL] += fil_size * sum(min_hops)
+        nhops += fil_size * sum(min_hops)
 
-    # Ifmap access.
+    return nhops
+
+
+@fastcache.clru_cache(maxsize=1024)
+def _unit_nhops_to_ifm(ifmap_layout, ifm_dict):
+    '''
+    Get the total number of hops to transfer ifmap data.
+
+    `ifm_dict` maps each fmap range to the nodes that process it.
+    '''
+    nhops = 0
 
     for ifrng, coord_list in ifm_dict.items():
-        nhops[de.IFM] += ifmap_layout.nhops_to(ifrng, *coord_list)
+        nhops += ifmap_layout.nhops_to(ifrng, *coord_list)
 
-    # Ofmap access.
+    return nhops
+
+
+@fastcache.clru_cache(maxsize=1024)
+def _unit_nhops_to_ofm(ofmap_layout, ofm_dict):
+    '''
+    Get the total number of hops to transfer ofmap data.
+
+    `ofm_dict` maps each fmap range to the nodes that process it.
+    '''
+    nhops = 0
 
     for ofrng, coord_list in ofm_dict.items():
         # Additional synchronization is necessary between INPP nodes. Only one
@@ -316,11 +362,11 @@ def unit_nhops_to_proc_region(layer, batch_size, region, part,
         for idx, coord in enumerate(coord_list):
             if idx == mid_idx:
                 # The mid node. Fetch from memory.
-                nhops[de.OFM] += ofmap_layout.nhops_to(ofrng, coord)
+                nhops += ofmap_layout.nhops_to(ofrng, coord)
             else:
                 # Others. Send to the mid node (one way).
                 dist = coord.hop_dist(coord_list[mid_idx])
-                nhops[de.OFM] += util.idivc(ofrng.size() * dist, 2)
+                nhops += util.idivc(ofrng.size() * dist, 2)
 
     return nhops