+++ nn_dataflow/tests/unit_test/test_resource.py
@@ -30,6 +30,8 @@ class TestResource(unittest.TestCase):
     def setUp(self):
         self.proc_region = NodeRegion(dim=PhyDim2(2, 2), origin=PhyDim2(0, 0),
                                       type=NodeRegion.PROC)
+        self.dram_region = NodeRegion(dim=PhyDim2(2, 2), origin=PhyDim2(0, 0),
+                                      type=NodeRegion.DRAM)
         self.src_data_region = NodeRegion(dim=PhyDim2(2, 1),
                                           origin=PhyDim2(0, 0),
                                           type=NodeRegion.DRAM)
@@ -40,6 +42,7 @@ class TestResource(unittest.TestCase):
     def test_valid_args(self):
         ''' Valid arguments. '''
         resource = Resource(proc_region=self.proc_region,
+                            dram_region=self.dram_region,
                             src_data_region=self.src_data_region,
                             dst_data_region=self.dst_data_region,
                             dim_array=PhyDim2(16, 16),
@@ -49,6 +52,7 @@ class TestResource(unittest.TestCase):
                             dram_bandwidth=128,
                            )
         self.assertTupleEqual(resource.proc_region.dim, (2, 2), 'proc_region')
+        self.assertTupleEqual(resource.dram_region.dim, (2, 2), 'dram_region')
         self.assertTupleEqual(resource.dim_array, (16, 16), 'dim_array')
         self.assertEqual(resource.size_gbuf, 131072, 'size_gbuf')
         self.assertEqual(resource.size_regf, 512, 'size_regf')
@@ -59,6 +63,7 @@ class TestResource(unittest.TestCase):
         ''' Invalid proc_region. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*proc_region.*'):
             _ = Resource(proc_region=PhyDim2(2, 2),
+                         dram_region=self.dram_region,
                          src_data_region=self.src_data_region,
                          dst_data_region=self.dst_data_region,
                          dim_array=PhyDim2(16, 16),
@@ -68,12 +73,43 @@ class TestResource(unittest.TestCase):
                          dram_bandwidth=128,
                         )
 
-    def test_invalid_proc_region_data(self):
+    def test_invalid_proc_region_dram(self):
         ''' Invalid proc_region with type DRAM. '''
         with self.assertRaisesRegexp(ValueError, 'Resource: .*proc_.*type.*'):
             _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
                                                 origin=PhyDim2(0, 0),
                                                 type=NodeRegion.DRAM),
+                         dram_region=self.dram_region,
+                         src_data_region=self.src_data_region,
+                         dst_data_region=self.dst_data_region,
+                         dim_array=PhyDim2(16, 16),
+                         size_gbuf=131072,
+                         size_regf=512,
+                         array_bus_width=8,
+                         dram_bandwidth=128,
+                        )
+
+    def test_invalid_dram_region(self):
+        ''' Invalid dram_region. '''
+        with self.assertRaisesRegexp(TypeError, 'Resource: .*dram_region.*'):
+            _ = Resource(proc_region=self.proc_region,
+                         dram_region=PhyDim2(2, 2),
+                         src_data_region=self.src_data_region,
+                         dst_data_region=self.dst_data_region,
+                         dim_array=PhyDim2(16, 16),
+                         size_gbuf=131072,
+                         size_regf=512,
+                         array_bus_width=8,
+                         dram_bandwidth=128,
+                        )
+
+    def test_invalid_dram_region_proc(self):
+        ''' Invalid dram_region with type DRAM. '''
+        with self.assertRaisesRegexp(ValueError, 'Resource: .*dram_.*type.*'):
+            _ = Resource(proc_region=self.proc_region,
+                         dram_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
                          src_data_region=self.src_data_region,
                          dst_data_region=self.dst_data_region,
                          dim_array=PhyDim2(16, 16),
@@ -87,6 +123,7 @@ class TestResource(unittest.TestCase):
         ''' Invalid src/dst_proc_region. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*src_data_.*'):
             _ = Resource(proc_region=self.proc_region,
+                         dram_region=self.dram_region,
                          src_data_region=PhyDim2(2, 1),
                          dst_data_region=self.dst_data_region,
                          dim_array=PhyDim2(16, 16),
@@ -97,6 +134,7 @@ class TestResource(unittest.TestCase):
                         )
         with self.assertRaisesRegexp(TypeError, 'Resource: .*dst_data_.*'):
             _ = Resource(proc_region=self.proc_region,
+                         dram_region=self.dram_region,
                          src_data_region=self.src_data_region,
                          dst_data_region=PhyDim2(2, 1),
                          dim_array=PhyDim2(16, 16),
@@ -110,6 +148,7 @@ class TestResource(unittest.TestCase):
         ''' Invalid dim_array. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*dim_array.*'):
             _ = Resource(proc_region=self.proc_region,
+                         dram_region=self.dram_region,
                          src_data_region=self.src_data_region,
                          dst_data_region=self.dst_data_region,
                          dim_array=(16, 16),
@@ -123,6 +162,7 @@ class TestResource(unittest.TestCase):
         ''' Invalid size_gbuf. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*size_gbuf.*'):
             _ = Resource(proc_region=self.proc_region,
+                         dram_region=self.dram_region,
                          src_data_region=self.src_data_region,
                          dst_data_region=self.dst_data_region,
                          dim_array=PhyDim2(16, 16),
@@ -136,6 +176,7 @@ class TestResource(unittest.TestCase):
         ''' Invalid size_regf. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*size_regf.*'):
             _ = Resource(proc_region=self.proc_region,
+                         dram_region=self.dram_region,
                          src_data_region=self.src_data_region,
                          dst_data_region=self.dst_data_region,
                          dim_array=PhyDim2(16, 16),
@@ -150,6 +191,7 @@ class TestResource(unittest.TestCase):
         with self.assertRaisesRegexp(TypeError,
                                      'Resource: .*array_bus_width.*'):
             _ = Resource(proc_region=self.proc_region,
+                         dram_region=self.dram_region,
                          src_data_region=self.src_data_region,
                          dst_data_region=self.dst_data_region,
                          dim_array=PhyDim2(16, 16),
@@ -161,6 +203,7 @@ class TestResource(unittest.TestCase):
         with self.assertRaisesRegexp(ValueError,
                                      'Resource: .*array_bus_width.*'):
             _ = Resource(proc_region=self.proc_region,
+                         dram_region=self.dram_region,
                          src_data_region=self.src_data_region,
                          dst_data_region=self.dst_data_region,
                          dim_array=PhyDim2(16, 16),
@@ -172,6 +215,7 @@ class TestResource(unittest.TestCase):
         with self.assertRaisesRegexp(ValueError,
                                      'Resource: .*array_bus_width.*'):
             _ = Resource(proc_region=self.proc_region,
+                         dram_region=self.dram_region,
                          src_data_region=self.src_data_region,
                          dst_data_region=self.dst_data_region,
                          dim_array=PhyDim2(16, 16),
@@ -186,6 +230,7 @@ class TestResource(unittest.TestCase):
         with self.assertRaisesRegexp(TypeError,
                                      'Resource: .*dram_bandwidth.*'):
             _ = Resource(proc_region=self.proc_region,
+                         dram_region=self.dram_region,
                          src_data_region=self.src_data_region,
                          dst_data_region=self.dst_data_region,
                          dim_array=PhyDim2(16, 16),
@@ -197,6 +242,7 @@ class TestResource(unittest.TestCase):
         with self.assertRaisesRegexp(ValueError,
                                      'Resource: .*dram_bandwidth.*'):
             _ = Resource(proc_region=self.proc_region,
+                         dram_region=self.dram_region,
                          src_data_region=self.src_data_region,
                          dst_data_region=self.dst_data_region,
                          dim_array=PhyDim2(16, 16),
@@ -208,6 +254,7 @@ class TestResource(unittest.TestCase):
         with self.assertRaisesRegexp(ValueError,
                                      'Resource: .*dram_bandwidth.*'):
             _ = Resource(proc_region=self.proc_region,
+                         dram_region=self.dram_region,
                          src_data_region=self.src_data_region,
                          dst_data_region=self.dst_data_region,
                          dim_array=PhyDim2(16, 16),