+++ CHANGELOG.md
@@ -51,7 +51,7 @@ List of major changes and improvements
   conservative type.
 
 
-## [v1.5 -- v1.6] -- 2018-07-12
+## [v1.5 -- v1.6] -- 2018-07-26
 
 ### Added
 
@@ -68,6 +68,8 @@ List of major changes and improvements
 
 - Hardware models:
 
+  - Add DRAM region in `Resource`.
+
   - Consider array bus width and its impact on data multicast latency.
 
   - Consider DRAM access time due to bandwidth limit.
@@ -80,6 +82,8 @@ List of major changes and improvements
     along one dimension unless it is the only partitioning.
 
   - Add utility `IntRange` for integer ranges.
+
+  - Add `HashableDict` class.
 
 
 ### Changed
@@ -141,6 +145,8 @@ List of major changes and improvements
 - ITCN access calculation for unit pass in `MapStrategy`.
 
 - `FmapRange` comparison.
+
+- Unit nhops calculation for filter data uses DRAM region.
 
 
 ## [v1.4 -- v1.5] -- 2017-08-18